+++ device_backends/NumericAddressedBackend/src/FixedPointConverter.cc
@@ -49,6 +49,7 @@ namespace ChimeraTK {
     // compute mask for the signed bit
     // keep the mask at 0 if unsigned to simplify further calculations
     if(nBits > 0) {
+      // NOLINTNEXTLINE(hicpp-signed-bitwise)
       _signBitMask = (_isSigned ? 1 << (nBits - 1) : 0x0); // the highest valid bit is the sign
     }
     else {
@@ -56,18 +57,24 @@ namespace ChimeraTK {
     }
 
     // compute masks of used and unused bits
-    _usedBitsMask = (1L << nBits) - 1L; // by using 1L (64 bit) this also works for 32 bits
+    // by using 1L (64 bit) this also works for 32 bits, which needs 33 bits during the calculation
+    // NOLINTNEXTLINE(hicpp-signed-bitwise)
+    _usedBitsMask = static_cast<int32_t>((1L << nBits) - 1L);
+    // NOLINTNEXTLINE(hicpp-signed-bitwise)
     _unusedBitsMask = ~_usedBitsMask;
 
     // compute bit shift mask, used to test if bit shifting for fractional bits
     // leads to an overflow
-    _bitShiftMask = ~(0xFFFFFFFF >> abs(_fractionalBits));
+    // NOLINTNEXTLINE(hicpp-signed-bitwise)
+    _bitShiftMask = static_cast<int32_t>(~(0xFFFFFFFF >> abs(_fractionalBits)));
 
     // compute minimum and maximum value in raw representation
+    // NOLINTNEXTLINE(hicpp-signed-bitwise)
     _maxRawValue = _usedBitsMask ^ _signBitMask; // bitwise xor: first bit is 0 if signed
+                                                 // NOLINTNEXTLINE(hicpp-signed-bitwise)
     _minRawValue = _signBitMask;                 // if only the sign bit is on, it is the smallest
                                                  // possible value
-    // (0 if unsigned)
+                                                 // (0 if unsigned)
 
     // fill all user type depending values: min and max cooked values and
     // fractional bit coefficients note: we loop over one of the maps only, but
@@ -78,6 +85,8 @@ namespace ChimeraTK {
   /**********************************************************************************************************************/
 
   template<>
+  // sorry, linter. We can't change the signature here. This is a template specialisation for std::string.
+  // NOLINTNEXTLINE(performance-unnecessary-value-param)
   uint32_t FixedPointConverter::toRaw<std::string>(std::string cookedValue) const {
     if(_fractionalBits == 0) { // use integer conversion
       if(_isSigned) {
